moved functions from orchestrator.py -> utils/*.py
